fix cast

+++ src/KubernetesClient/LeaderElection/LeaderElector.cs
@@ -111,7 +111,7 @@ namespace k8s.LeaderElection
             var leaderElectionRecord = new LeaderElectionRecord()
             {
                 HolderIdentity = l.Identity,
-                LeaseDurationSeconds = config.LeaseDuration.TotalSeconds,
+                LeaseDurationSeconds = (int)config.LeaseDuration.TotalSeconds,
                 AcquireTime = DateTime.UtcNow,
                 RenewTime = DateTime.UtcNow,
                 LeaderTransitions = 0,
